Fix blockquote nesting, move tinymce.css to assets
