+++ app/assets/stylesheets/layouts/starrydark.scss
@@ -1,6 +1,7 @@
-﻿#header { background-image: url("/images/starry.jpg"); background-position: right top; }
+@import 'variables';
+#header { background-image: url("/images/starry.jpg"); background-position: right top; }
 #logo, #nav-bottom { background-color: transparent !important; }
-#header #user-info, #header #nav-top { background-color: #05050F; background-color: rgba(5, 5, 15, 0.9); }
+#user-info, #nav-top { background-color: #05050F; background-color: rgba(5, 5, 15, 0.9); }
 
 $FONT_COLOR: #DADADA;
 $FONT_COLOR_DARK: #444444;
@@ -16,11 +17,33 @@ th, .single-description, .content-header {
   a:visited { color: #9C9AA4; }
 }
 
-#post-menu-box { border-top: 1px solid #808080; border-top: 1px solid rgba(128,128,128,0.5); }
-#post-menu-box div { border-bottom: 1px solid #808080; border-bottom: 1px solid rgba(128,128,128,0.5); }
-#post-menu-box a { color: inherit; }
-#content .even, .post-post, .post-reply:nth-child(even), #post-editor, .odd blockquote, .post-reply:nth-child(odd) blockquote { background-color: #31323B; }
-#content .odd, .post-reply:nth-child(odd), .post-post + #post-editor, .post-post + a + #post-editor, .even blockquote, .post-post blockquote, .post-reply:nth-child(even) blockquote { background-color: #25262E; }
+$EVEN_COLOR: #31323B;
+$ODD_COLOR: #25262E;
+
+#post-menu-box {
+  border-top: 1px solid #808080;
+  border-top: 1px solid rgba(128,128,128,0.5);
+
+  a { color: inherit; }
+  div {
+    border-bottom: 1px solid #808080;
+    border-bottom: 1px solid rgba(128,128,128,0.5);
+  }
+}
+.even, .post-post, .post-reply:nth-child(even), #post-editor {
+  background-color: $EVEN_COLOR;
+
+  blockquote {
+    background-color: calculate-transparent-color($EVEN_COLOR, $ODD_COLOR, 0.2);
+  }
+}
+.odd, .post-reply:nth-child(odd), .post-post + #post-editor, .post-post + a + #post-editor {
+  background-color: $ODD_COLOR;
+
+  blockquote {
+    background-color: calculate-transparent-color($ODD_COLOR, $EVEN_COLOR, 0.2);
+  }
+}
 
 .post-navheader .view-button, .view-button.selected, .view-button, .link-box.green, .link-box.blue, .link-box.red, .tag-item { background-color: #463C65; }
 .flash.breadcrumbs .link-box, .subber .link-box { background-color: #5C5778; }
