+++ app/assets/stylesheets/tinymce.scss
@@ -1,0 +1,33 @@
+@import 'variables';
+body {
+  font-family: Helvetica, sans-serif;
+  font-size: 14px;
+  background-color: $bg_color_input;
+
+  blockquote {
+    background-color: calculate-transparent-color($bg_color_input, #EAEAEA, 0.2);
+    margin: 0.5em 1.5em;
+    padding: 0.5em;
+
+    & > p:first-child { margin-top: 0; }
+    & > p:last-child { margin-bottom: 0; }
+  }
+
+  &.layout_starrydark {
+    background-color: #4D4B5B;
+    color: #DADADA;
+
+    blockquote {
+      background-color: calculate-transparent-color(#4D4B5B, #31323B, 0.12);
+    }
+  }
+
+  &.layout_dark {
+    background-color: #4D4D4D;
+    color: #DADADA;
+
+    blockquote {
+      background-color: calculate-transparent-color(#4D4D4D, #3E3E3E, 0.12);
+    }
+  }
+}
