+++ config/initializers/assets.rb
@@ -9,4 +9,4 @@ Rails.application.config.assets.version = '1.0'
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
 Rails.application.config.assets.precompile += %w( posts/*.js users/*.js boards/*.js characters/*.js global.js paginator.js galleries/*.js messages.js icons.js board_sections.js icons.js reorder.js )
-Rails.application.config.assets.precompile += %w( layouts/*.css )
+Rails.application.config.assets.precompile += %w( layouts/*.css tinymce.css )
