--- public/stylesheets/tinymce.css
@@ -1,19 +1,0 @@
-body {
-  font-family: Helvetica, sans-serif;
-  font-size: 14px;
-}
-body.layout_starrydark {
-  background-color: #4D4B5B;
-  color: #DADADA;
-}
-body.layout_dark {
-  background-color: #4D4D4D;
-  color: #DADADA;
-}
-blockquote {
-  background-color: #E3E3E3;
-  background-color: rgba(128, 128, 128, 0.2);
-  margin: 0.5em 1.5em; padding: 0.5em;
-}
-blockquote > p:first-child { margin-top: 0; }
-blockquote > p:last-child { margin-bottom: 0; }
